Loader: Split SMDH into its own header and import helpers from QGameList
Also rewrite Qt wrappers to use those.
