+++ src/core/loader/loader.h
@@ -10,10 +10,8 @@
 #include <string>
 #include <vector>
 
-#include "common/common_funcs.h"
 #include "common/common_types.h"
 #include "common/file_util.h"
-#include "common/swap.h"
 
 namespace Kernel {
 struct AddressMapping;
@@ -79,51 +77,6 @@ enum class ResultStatus {
 constexpr u32 MakeMagic(char a, char b, char c, char d) {
     return a | b << 8 | c << 16 | d << 24;
 }
-
-/// SMDH data structure that contains titles, icons etc. See https://www.3dbrew.org/wiki/SMDH
-struct SMDH {
-    u32_le magic;
-    u16_le version;
-    INSERT_PADDING_BYTES(2);
-
-    struct Title {
-        std::array<u16, 0x40> short_title;
-        std::array<u16, 0x80> long_title;
-        std::array<u16, 0x40> publisher;
-    };
-    std::array<Title, 16> titles;
-
-    std::array<u8, 16> ratings;
-    u32_le region_lockout;
-    u32_le match_maker_id;
-    u64_le match_maker_bit_id;
-    u32_le flags;
-    u16_le eula_version;
-    INSERT_PADDING_BYTES(2);
-    float_le banner_animation_frame;
-    u32_le cec_id;
-    INSERT_PADDING_BYTES(8);
-
-    std::array<u8, 0x480> small_icon;
-    std::array<u8, 0x1200> large_icon;
-
-    /// indicates the language used for each title entry
-    enum class TitleLanguage {
-        Japanese = 0,
-        English = 1,
-        French = 2,
-        German = 3,
-        Italian = 4,
-        Spanish = 5,
-        SimplifiedChinese = 6,
-        Korean= 7,
-        Dutch = 8,
-        Portuguese = 9,
-        Russian = 10,
-        TraditionalChinese = 11
-    };
-};
-static_assert(sizeof(SMDH) == 0x36C0, "SMDH structure size is wrong");
 
 /// Interface for loading an application
 class AppLoader : NonCopyable {
